created card component and refactored Products.js
